# chore: Experimenting with Oxlint & Biome