--- a/web/.prettierignore
+++ b/web/.prettierignore
@@ -13,3 +13,5 @@ node_modules
 pnpm-lock.yaml
 package-lock.json
 yarn.lock
+biome.json
+.oxlintrc.json