--- a/web/eslint.config.js
+++ b/web/eslint.config.js
@@ -99,5 +99,4 @@ export default typescriptEslint.config(
       },
     },
   },
-  oxlint.buildFromOxlintConfigFile('./.oxlintrc.json'),
 );