# What is this?
We (@danieldietzler & me) wanted to try out the newly released 🦀-based linter  [Oxlint](https://oxc.rs/blog/2025-06-10-oxlint-stable.html) so see how feasible a migration from eslint to Oxlint would be, and what performance enhancement could be gained.

I also explored how [BiomeJS](https://biomejs.dev/) would fare.

Also since the main pain-point is web, I'm only taking a look at that, but the server could ofc also benefit in the same way (with probably less migration problems).

# Migration
Both Oxlint and Biome offer a migration tool to migrate from eslint to their respective tool. That is needed, since both tools (re-)implement their own rules, rules written for eslint cannot simply be used.

## Oxlint
For the oxlint migration, it helpfully outputs which rules are not supported (yet) which gives a great overview:

<details><summary>Oxlint migration output</summary>

```
unsupported rule: svelte/comment-directive
unsupported rule: svelte/system
special parser detected: svelte-eslint-parser
special parser detected: svelte-eslint-parser
unsupported rule: svelte/comment-directive
unsupported rule: svelte/infinite-reactive-loop
unsupported rule: svelte/no-at-debug-tags
unsupported rule: svelte/no-at-html-tags
unsupported rule: svelte/no-dom-manipulating
unsupported rule: svelte/no-dupe-else-if-blocks
unsupported rule: svelte/no-dupe-on-directives
unsupported rule: svelte/no-dupe-style-properties
unsupported rule: svelte/no-dupe-use-directives
unsupported rule: svelte/no-export-load-in-svelte-module-in-kit-pages
unsupported rule: svelte/no-immutable-reactive-statements
unsupported rule: svelte/no-inner-declarations
unsupported rule: svelte/no-inspect
unsupported rule: svelte/no-not-function-handler
unsupported rule: svelte/no-object-in-text-mustaches
unsupported rule: svelte/no-raw-special-elements
unsupported rule: svelte/no-reactive-functions
unsupported rule: svelte/no-reactive-literals
unsupported rule: svelte/no-reactive-reassign
unsupported rule: svelte/no-shorthand-style-property-overrides
unsupported rule: svelte/no-store-async
unsupported rule: svelte/no-svelte-internal
unsupported rule: svelte/no-unknown-style-directive-property
unsupported rule: svelte/no-unnecessary-state-wrap
unsupported rule: svelte/no-unused-props
unsupported rule: svelte/no-unused-svelte-ignore
unsupported rule: svelte/no-useless-children-snippet
unsupported rule: svelte/no-useless-mustaches
unsupported rule: svelte/prefer-writable-derived
unsupported rule: svelte/require-each-key
unsupported rule: svelte/require-event-dispatcher-types
unsupported rule: svelte/require-store-reactive-access
unsupported rule: svelte/system
unsupported rule: svelte/valid-each-key
unsupported rule: svelte/valid-prop-names-in-kit-pages
unsupported rule: unicorn/expiring-todo-comments
unsupported rule: unicorn/import-style
unsupported rule: unicorn/no-array-callback-reference
unsupported rule: unicorn/no-for-loop
unsupported rule: unicorn/no-named-default
unsupported rule: unicorn/no-unnecessary-array-splice-count
unsupported rule: unicorn/no-unnecessary-polyfills
unsupported rule: unicorn/prefer-at
unsupported rule: unicorn/prefer-default-parameters
unsupported rule: unicorn/prefer-export-from
unsupported rule: unicorn/prefer-keyboard-event-key
unsupported rule: unicorn/prefer-module
unsupported rule: unicorn/prefer-single-call
unsupported rule: unicorn/prefer-switch
unsupported rule: unicorn/prefer-ternary
unsupported rule: unicorn/prefer-top-level-await
unsupported rule: unicorn/prevent-abbreviations
unsupported rule: unicorn/relative-url-style
unsupported rule: unicorn/template-indent
unsupported rule: constructor-super
unsupported rule, but in development: getter-return
unsupported rule: no-dupe-args
unsupported rule: no-misleading-character-class
unsupported rule: no-octal
unsupported rule, but in development: no-undef
unsupported rule, but in development: no-unreachable
ignore allow list is currently not supported: !\*\*/.env.example
unsupported rule, but in development: getter-return
unsupported rule, but in development: no-undef
unsupported rule, but in development: no-unreachable
unsupported rule: prefer-const
unsupported rule: svelte/button-has-type
unsupported rule: @typescript-eslint/await-thenable
unsupported rule: @typescript-eslint/no-floating-promises
unsupported rule: @typescript-eslint/no-misused-promises
unsupported rule: @typescript-eslint/require-await
unsupported rule: object-shorthand
special parser detected: svelte-eslint-parser
```
</details> 

To summarize:
- type aware linting is not supported (https://github.com/oxc-project/oxc/issues/3105)
- svelte linting rules are also currently not supported, though there might potentially be work on the way over here? https://github.com/oxc-project/oxc/issues/11495#issuecomment-2956051716

## Biome
For Biome unfortunately things are much more unclear. The migration does not give out any feedback on which rules are unsupported and thus not migrated. So this would be tedious manual work, to find all the rules that are not supported.

What we do know, is that it also does not support type-aware linting (yet) (https://github.com/biomejs/biome/issues/3187)

And after running a first lint, I found biome flagged a lot of unused variable errors in svelte files, for variables that were only used in the svelte markup

# Speed
Now even with a significant amount of rules not working in both new tools, looking at those juicy timing numbers is still very fun, so here are some *very* un-scientific numbers of just a few one-shot runs I did on my PC (which has 16 threads)

## Current state
Single threaded:
`npm run lint` -> 516s (8min 16s)

Parallel:
`npm run lint:p` -> 174s (3min 14s)

## Oxlint
`npm run lint:ox` -> 45ms

## Biome
`npx @biomejs/biome lint` -> 87ms

# Further remarks on Oxlint
Oxlint also has a very interesting solution for rules that have not yet been implemented: [eslint-plugin-oxlint](https://github.com/oxc-project/eslint-plugin-oxlint)

This is an *eslint* plugin, that disabled all eslint rules, that are now covered by Oxlint. Thus you can then run as your lint command:
```
npx oxlint && npx eslint
```

However, I either didn't really manage to include this plugin correctly, or it made no significant change to the runtime of eslint. Ofc the main rules causing the slowness are the type-aware promise rules, which would fallback to eslint in this configuration. But even with those disabled and supposedly all the Oxlint rules disabled with the plugin, the eslint run time was still at least 50s which I find odd.

# Conclusion
From my experimenting, Oxlint looks like a safer bet, even though biome has been around for longer. With Oxlint we know exactly which rules are still missing which is a *huge* plus imo.

The main blocker for now would be the svelte lint support.
Given how atrociously slow the type-aware rules currently are, and how there has already been discussions to possibly remove them (in some circumstances), I feel like those aren't that big of a blocker.

## Notes
Both tools do also offer a formatter, which might be worth it to migrate to, if we ever migrate the linter, just for some extra 🦀