@arendjr 2025-06-13
Hi, Biome contributor popping in here :)

I just wanted to let you know that we are launching Biome 2 shortly and will be introducing type-aware linting with rules such as [`noFloatingPromises`](https://next.biomejs.dev/linter/rules/no-floating-promises/) and [`useExhaustiveSwitchCases`](https://next.biomejs.dev/linter/rules/use-exhaustive-switch-cases/). Others are already on the way.

The point about the migration not listing unsupported rules is a great point of feedback and something we intend to address! In the meantime, you can use our page with [rule sources](https://biomejs.dev/linter/rules-sources/) to look up rules that you feel are important to you.

Cheers!