@arnolicious 2025-06-13
Unfortunately the svelte files are the main bottleneck, so switching the non-svelte files to biome would not really be worth it imo.

For comparison, running the current eslint-parallel config takes around 166s, and (very crudely) stripping all svelte plugins and configs from the eslint config, results in barely 5s.

So as long as I didn't make some blunder, it does seem like the main problem are the svelte files