@benmccann 2025-06-13
I guess my comment was a bit unclear. I mean could we have the eslint run the `eslint-plugin-svelte` rules? Those are quite fast. It's applying the type-checked rules from `estlint-typescript` to Svelte files that is slow