@arnolicious 2025-06-13
Oh right, yeah that's a good idea, that should be fairly straight-forward!